+++ changelog.md
@@ -16,6 +16,7 @@
 * Improve error message when a query is interrupted by a signal (SIGINT or SIGTERM).
 * Improve error message if session is expired.
 * Improve robustness of Arrow client against connection errors such as `FlightUnavailableError` and `FlightTimedOutError`.
+* Return dedicated error class `SessionStatusError` if a session failed or expired.
 
 
 ## Other changes